Back-End/app/requirements.txt
flask
pymongo
pika
python-dotenv
